+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# WIP
+* i : Move from Mustache templates to Twig using [twig.js](https://github.com/twigjs/twig.js)
+* i : Add DNSlytics link with Page IP / Status display
+* \+ : Native messaging - accept showing notifications from native messaging
+
 # 1.0.0
 * i : Move to manifest v3, folders rework (remove data folder)
 * i : WIP native messaging (timeout system, getPreferences result, theme option renaming)